remove warning

+++ AlicloudHttpDNSTests/XCTestCase+AsyncTesting.m
@@ -107,7 +107,7 @@ static void *kExpectedStatus_Key = "kExpectedStatus_Key";
 
 - (void) setNotifiedStatus:(XCTAsyncTestCaseStatus)value
 {
-    NSNumber *valueNumber = [NSNumber numberWithInt:value];
+    NSNumber *valueNumber = [NSNumber numberWithInt:(int)value];
     [self setAssociatedObject:valueNumber key:kNotifiedStatus_Key];
 }
 
@@ -119,7 +119,7 @@ static void *kExpectedStatus_Key = "kExpectedStatus_Key";
 
 - (void) setExpectedStatus:(XCTAsyncTestCaseStatus)value
 {
-    NSNumber *valueNumber = [NSNumber numberWithInt:value];
+    NSNumber *valueNumber = [NSNumber numberWithInt:(int)value];
     [self setAssociatedObject:valueNumber key:kExpectedStatus_Key];
 }
